MIPS: add board support for AVM FritzBox 3370
There is only 128K for the bootloader and no space left for the
environment. To fit into the space available only a stripped down
version of u-boot can be used with no permanent environment.

So far only the EVA second stage loader was tested and only with Hynix
NAND. Remaining issue is a not working ethernet when started from EVA.

Signed-off-by: Daniel Schwierzeck <daniel.schwierzeck@gmail.com>