do not append the gphy firmware to the uboot image
The gphy firmware is embedded into to the c code since cd0eca3. Appending
the firmware once more to the u-boot binary is unnecessary. And in combination
with the the padding - which was used to have the firmware blob at a predictable
address - the u-boot image size exceeds 128Kb, which is the u-boot partition
size of the TDW8970.