MIPS: fixup board support for BT HomeHub 5A
Revert to ddr_settings.h from the stock u-boot GPL tarball. The
previous values are the ones optimized for NAND boot (with memory
tuning enabled). These might not work with UART boot.

According to mailing list posts, the recommend way to take care of bad
blocks in the env space is to use an env range instead of a redundant
env.

Drop the CONFIG_ENV_UPDATE_UBOOT_NAND to keep the env clean.

Add UBI support and load the kernel from the ubi volume kernel.

Set the GPIO value of the power led only once.

Signed-off-by: Mathias Kresin <dev@kresin.me>